# openapi.yaml
openapi: 3.0.3
info:
  title: FPL Assistant API
  description: >
    Backend API for the FPL Assistant website and MCP server (see
    PRD-fpl-assistant.md and TDD-fpl-assistant.md). Covers team lookup, player
    stats/search/comparison, fixtures (with double/blank gameweek detection),
    squad view, captain/transfer suggestions, mini-league standings and rival
    comparison, and chart data.

    Draft v0.1 — pending review (see tasks/todo.md Task 4). Not yet locked.
  version: 1.5.0
  contact:
    name: FPL Assistant
    email: jeffvu304@gmail.com
servers:
  - url: https://api.fplassistant.example.com
    description: Production (placeholder host, per TDD's deferred hosting decision)
...
paths:
...
  /teams/{teamId}/squad:
    get:
      operationId: getSquad
      tags: [Team]
      summary: Get a team's current squad (pitch/formation view)
      description: Returns the 15-player squad, current formation, and captain/vice-captain picks.
      parameters:
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Squad found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Squad'
              example:
                team_id: 1234567
                formation: "3-4-3"
                picks:
                  - player_id: 302
                    name: "M. Salah"
                    position: "MID"
                    is_captain: true
                    is_vice_captain: false
                  - player_id: 401
                    name: "E. Haaland"
                    position: "FWD"
                    is_captain: false
                    is_vice_captain: true
                data_as_of: "2026-09-06T09:00:00Z"
                stale: false
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /players:
    get:
      operationId: searchPlayers
      tags: [Players]
      summary: Search players
      description: Free-text and filtered player search, e.g. for building a comparison or scouting a transfer target.
      parameters:
        - name: query
          in: query
          description: Free-text search on player name
          schema:
            type: string
        - name: position
          in: query
          description: Filter by position
          schema:
            type: string
            enum: [GKP, DEF, MID, FWD]
        - name: team
          in: query
          description: Filter by real-world team name or short code
          schema:
            type: string
      responses:
        '200':
          description: Matching players
          content:
            application/json:
              schema:
                type: object
                properties:
                  players:
                    type: array
                    items:
                      $ref: '#/components/schemas/Player'
                  data_as_of:
                    $ref: '#/components/schemas/DataAsOf'
                  stale:
                    $ref: '#/components/schemas/Stale'
              example:
                players:
                  - player_id: 302
                    name: "M. Salah"
                    team: "Liverpool"
                    position: "MID"
                    price: 12.8
                    total_points: 68
                data_as_of: "2026-09-06T09:00:00Z"
                stale: false
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /players/{playerId}:
    get:
      operationId: getPlayerStats
      tags: [Players]
      summary: Get a single player's stats
      description: Returns price, points, form, and other stats for one player.
      parameters:
        - $ref: '#/components/parameters/PlayerId'
      responses:
        '200':
          description: Player found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
              example:
                player_id: 302
                name: "M. Salah"
                team: "Liverpool"
                position: "MID"
                price: 12.8
                total_points: 68
                form: 6.4
                data_as_of: "2026-09-06T09:00:00Z"
                stale: false
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /players/compare:
    get:
      operationId: comparePlayers
      tags: [Players]
      summary: Side-by-side comparison of two or more players
      description: Returns full stats for each requested player so they can be compared side by side.
      parameters:
        - name: playerIds
          in: query
          required: true
          description: Comma-separated player IDs to compare (2 or more)
          schema:
            type: string
          example: "302,401"
      responses:
        '200':
          description: Comparison result
          content:
            application/json:
              schema:
                type: object
                properties:
                  players:
                    type: array
                    items:
                      $ref: '#/components/schemas/Player'
                  data_as_of:
                    $ref: '#/components/schemas/DataAsOf'
                  stale:
                    $ref: '#/components/schemas/Stale'
              example:
                players:
                  - player_id: 302
                    name: "M. Salah"
                    team: "Liverpool"
                    position: "MID"
                    price: 12.8
                    total_points: 68
                  - player_id: 401
                    name: "E. Haaland"
                    team: "Man City"
                    position: "FWD"
                    price: 15.1
                    total_points: 74
                data_as_of: "2026-09-06T09:00:00Z"
                stale: false
        '400':
          description: Fewer than two valid player IDs supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "At least two playerIds are required for comparison"
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /fixtures:
    get:
      operationId: getFixtures
      tags: [Fixtures]
      summary: Get fixture difficulty ratings
      description: >
        Includes is_double_gameweek / is_blank_gameweek flags per team per
        gameweek, per TDD-fpl-assistant.md's DGW/BGW detection requirement.
      parameters:
        - name: gameweek
          in: query
          description: Filter to a specific gameweek
          schema:
            type: integer
        - name: teamId
          in: query
          description: Filter to fixtures involving a specific real-world FPL team ID
          schema:
            type: integer
      responses:
        '200':
          description: Fixture list
          content:
            application/json:
              schema:
                type: object
                properties:
                  fixtures:
                    type: array
                    items:
                      $ref: '#/components/schemas/Fixture'
                  data_as_of:
                    $ref: '#/components/schemas/DataAsOf'
                  stale:
                    $ref: '#/components/schemas/Stale'
              example:
                fixtures:
                  - gameweek: 7
                    home_team: "Liverpool"
                    away_team: "Everton"
                    difficulty_home: 2
                    difficulty_away: 4
                    is_double_gameweek: false
                    is_blank_gameweek: false
                  - gameweek: 18
                    home_team: "Man City"
                    away_team: null
                    difficulty_home: null
                    difficulty_away: null
                    is_double_gameweek: false
                    is_blank_gameweek: true
                data_as_of: "2026-09-06T09:00:00Z"
                stale: false
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /teams/{teamId}/suggest-captain:
    get:
      operationId: suggestCaptain
      tags: [Recommendations]
      summary: Suggest a captain for the given team
      description: >
        Opinionated tool — returns the shared RecommendationContract shape
        (TDD-fpl-assistant.md worked example: `recommendation` is
        `{ player_id, name }`).
      parameters:
        - $ref: '#/components/parameters/TeamId'
        - name: gameweek
          in: query
          description: Gameweek to suggest a captain for (defaults to the next unfinished gameweek)
          schema:
            type: integer
      responses:
        '200':
          description: Captain recommendation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationContract'
              example:
                recommendation:
                  player_id: 302
                  name: "M. Salah"
                confidence: "high"
                reasoning:
                  - factor: "fixture_difficulty"
                    detail: "Faces the league's 18th-ranked defense at home"
                    sample_size: 12
                  - factor: "recent_form"
                    detail: "Scored or assisted in 5 of the last 6 gameweeks"
                    sample_size: 6
                alternatives:
                  - option:
                      player_id: 401
                      name: "E. Haaland"
                    reasoning:
                      - factor: "recent_form"
                        detail: "4 goals in the last 3 gameweeks"
                        sample_size: 3
                    why_not_top_pick: "Tougher away fixture against a top-6 defense"
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /mini-leagues/{leagueId}/standings:
    get:
      operationId: getMiniLeagueStandings
      tags: [MiniLeague]
      summary: Get mini-league standings
      description: Returns the current standings table for a public FPL mini-league.
      parameters:
        - $ref: '#/components/parameters/LeagueId'
      responses:
        '200':
          description: Standings found
          content:
            application/json:
              schema:
                type: object
                properties:
                  league_id:
                    type: integer
                  league_name:
                    type: string
                  standings:
                    type: array
                    items:
                      $ref: '#/components/schemas/MiniLeagueEntry'
                  data_as_of:
                    $ref: '#/components/schemas/DataAsOf'
                  stale:
                    $ref: '#/components/schemas/Stale'
              example:
                league_id: 987654
                league_name: "Office League"
                standings:
                  - rank: 1
                    team_id: 1234567
                    team_name: "Fantasy Wanderers"
                    total_points: 412
                  - rank: 2
                    team_id: 7654321
                    team_name: "Rival FC"
                    total_points: 398
                data_as_of: "2026-09-06T09:00:00Z"
                stale: false
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /teams/{teamId}/charts:
    get:
      operationId: getChartData
      tags: [Charts]
      summary: Get chart data for a team
      description: Returns time-series data points for points-per-gameweek, rank trajectory, or price history.
      parameters:
        - $ref: '#/components/parameters/TeamId'
        - name: type
          in: query
          required: true
          description: Which chart to fetch
          schema:
            type: string
            enum: [points_per_gameweek, rank_trajectory, price_history]
      responses:
        '200':
          description: Chart data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChartData'
              example:
                type: "points_per_gameweek"
                points:
                  - x: 1
                    y: 62
                  - x: 2
                    y: 78
                  - x: 3
                    y: 55
                data_as_of: "2026-09-06T09:00:00Z"
                stale: false
        '400':
          description: The required `type` query parameter is missing or not one of the allowed values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "type must be one of: points_per_gameweek, rank_trajectory, price_history"
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  parameters:
    TeamId:
      name: teamId
      in: path
      required: true
      description: Public FPL team ID
      schema:
        type: integer
    PlayerId:
      name: playerId
      in: path
      required: true
      schema:
        type: integer
    LeagueId:
      name: leagueId
      in: path
      required: true
      description: Public FPL mini-league ID
      schema:
        type: integer
  responses:
    NotFound:
      description: The requested ID doesn't exist (e.g. invalid/unknown team ID)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "That team ID doesn't exist"
    TooManyRequests:
      description: Rate limit exceeded (60 requests/minute/IP)
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: string
            pattern: '^[0-9]+$'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Rate limit exceeded, please retry shortly"
    ServiceUnavailable:
      description: Recommendation couldn't be produced because required data is unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DegradedState'
          example:
            message: "Unable to produce a recommendati
            on right now, please retry shortly"
  schemas:
    DataAsOf:
      type: string
      format: date-time
      description: Timestamp of the underlying cached data this response was served from
    Stale:
      type: boolean
      description: >
        True if this response was served from a cached value because the live
        FPL API was unreachable (graceful degradation, not an error state)
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
    DegradedState:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Squad:
      type: object
      properties:
        team_id:
          type: integer
        formation:
          type: string
        picks:
          type: array
          items:
            $ref: '#/components/schemas/SquadPick'
        data_as_of:
          $ref: '#/components/schemas/DataAsOf'
        stale:
          $ref: '#/components/schemas/Stale'
    SquadPick:
      type: object
      properties:
        player_id:
          type: integer
        name:
          type: string
        position:
          type: string
          enum: [GKP, DEF, MID, FWD]
        is_captain:
          type: boolean
        is_vice_captain:
          type: boolean
    Player:
      type: object
      properties:
        player_id:
          type: integer
        name:
          type: string
        team:
          type: string
        position:
          type: string
          enum: [GKP, DEF, MID, FWD]
        price:
          type: number
        total_points:
          type: integer
        form:
          type: number
        data_as_of:
          $ref: '#/components/schemas/DataAsOf'
        stale:
          $ref: '#/components/schemas/Stale'
    Fixture:
      type: object
      properties:
        gameweek:
          type: integer
        home_team:
          type: string
          nullable: true
        away_team:
          type: string
          nullable: true
        difficulty_home:
          type: integer
          nullable: true
        difficulty_away:
          type: integer
          nullable: true
        is_double_gameweek:
          type: boolean
          description: True if the team plays more than once in this gameweek
        is_blank_gameweek:
          type: boolean
          description: True if the team has no fixture in this gameweek
    MiniLeagueEntry:
      type: object
      properties:
        rank:
          type: integer
        team_id:
          type: integer
        team_name:
          type: string
        total_points:
          type: integer
    ChartData:
      type: object
      properties:
        type:
          type: string
          enum: [points_per_gameweek, rank_trajectory, price_history]
        points:
          type: array
          items:
            type: object
            properties:
              x:
                type: number
                description: Gameweek number (points_per_gameweek, rank_trajectory) or date (price_history)
              y:
                type: number
                description: Points, rank, or price value
        data_as_of:
          $ref: '#/components/schemas/DataAsOf'
        stale:
          $ref: '#/components/schemas/Stale'

    # Shared response shape for every opinionated tool (captain suggestion,
    # transfer suggestion, chip timing). Mirrors
    # api/app/models/recommendation_contract.rb exactly. `recommendation` and
    # `alternatives[].option` are intentionally free-form objects because
    # their internal shape differs per tool (see PORO comment / TDD worked
    # examples) — e.g. suggest_captain's recommendation is
    # { player_id, name }, suggest_transfer's is { player_out, player_in }.
    RecommendationContract:
      type: object
      required: [recommendation, confidence, reasoning, alternatives]
      properties:
        recommendation:
          type: object
          description: Shape varies per tool (see suggest_captain / suggest_transfer examples)
          additionalProperties: true
        confidence:
          type: string
          enum: [low, medium, high]
          description: >
            Derived from sample size of the relevant signal: fewer than 3
            relevant gameweeks -> low, 3-10 -> medium, 10+ -> high. Reflects
            signal strength only, never a claim of correctness.
        reasoning:
          type: array
          items:
            $ref: '#/components/schemas/ReasoningFactor'
        alternatives:
          type: array
          items:
            $ref: '#/components/schemas/Alternative'
        data_as_of:
          type: string
          format: date-time
          nullable: true
          description: >
            Present only when this recommendation was served from stale
            cached data; the timestamp the underlying data was last
            successfully fetched.
    ReasoningFactor:
      type: object
      required: [factor, detail, sample_size]
      properties:
        factor:
          type: string
        detail:
          type: string
        sample_size:
          type: integer
    Alternative:
      type: object
      required: [option, reasoning, why_not_top_pick]
      properties:
        option:
          type: object
          description: Shape matches the sibling `recommendation` field for the same tool
          additionalProperties: true
        reasoning:
          type: array
          items:
            $ref: '#/components/schemas/ReasoningFactor'
        why_not_top_pick:
          type: string